backend/requirements.txt
asgiref>=3.6.0,<4
Django==4.2.1
psycopg2-binary==2.9.5
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.2.0
gunicorn==20.1.0
python-dotenv==1.0.1
# django rest framework
djangorestframework==3.15.1
djangorestframework-simplejwt==4.8.0
django-cors-headers==3.11.0
# redis
django-redis==5.4.0
redis==5.0.7
# celery
celery==5.4.0
# utils
requests==2.26.0
slack-sdk==3.34.0